drop legacy Piper naming and extract shared download infra
Renames the last legacy "Piper"-branded internal identifiers in
addon/ now that Kokoro is a second backend, and pulls the download
machinery kokoro_download.py was already reaching into
voice_download.py's private names for out into a shared
download_infra.py.

- aio.py: renamed the executor/event-loop thread names from
  piper4nvda_* to dengjen_*.
- installTasks.py: renamed _PIPER_SYNTH_DIR to _SYNTH_DRIVER_DIR.
- domain/tts_system.py: fixed the stale "No Piper voices found"
  exception message (never surfaced to users -- reaches only the
  NVDA log).
- voice_manager.py: renamed OnlineDengjenVoicesPanel to
  OnlinePiperVoicesPanel, since it lists Piper's online catalog
  specifically and Kokoro now has its own tab; also fixed two stale
  references in model_catalog.py's docstring/error message.
- New download_infra.py: extracted the backend-agnostic
  redirect-following, TLS trust-store-gap fallback, resumable disk
  streaming, and the shared BaseVoiceDownloader workflow out of
  voice_download.py. voice_download.py keeps only Piper-specific
  catalog/download code; kokoro_download.py and voice_manager.py
  import the shared pieces from download_infra.py instead of
  reaching into voice_download.py's private names. This module's
  cross-module API (BaseVoiceDownloader, VoiceInstallError,
  follow_redirects, get_with_cert_fallback, stream_to_file,
  resumable_partial_size, archive_total_size) is public -- no
  leading underscore -- since that's exactly what these names are
  for; the three names nothing outside this file uses stay private.
- Test coverage for the moved functions moved to a new
  tests/test_download_infra.py; the tests remaining in
  tests/test_voice_download.py had their request/THREAD_POOL_EXECUTOR/
  DENGJEN_VOICES_DIR/gui/wx/core monkeypatches retargeted, since
  Python resolves a function's globals against its defining module.
- Two pre-existing download-flow bugs CodeRabbit flagged against
  this diff, fixed with regression tests: done_callback now catches
  Exception (not just VoiceInstallError) from _install, since
  PiperVoiceDownloader._install's unwrapped voice_dir.mkdir() could
  otherwise escape the worker-thread callback silently; and
  PiperRTVoiceDownloader._do_download_archive now discards a stale,
  already-complete leftover archive and retries once from scratch
  instead of failing on a Range request past EOF (416).

No persisted-state changes (on-disk voices/piper/ path, config keys)
and no other user-facing behavior changes.

Closes #139
